직전 commit에서 반영 안된것 recommit

+++ src/main/java/example/springboard/service/BoardService.java
@@ -14,5 +14,5 @@ public interface BoardService {
     Board writeBoard(Board board);
     int updateBoard(Board board);
     int deleteBoard(Board board);
-    int writeBoardReply(Long id, Board board);
+    Long writeBoardReply(Long id, Board board);
 }
